+++ src/lib/components/modal/list/CreateNewList.tsx
@@ -97,7 +97,7 @@ export function CreateNewListModal({
   return (
     <ActionModal
       title="Create a New List"
-      icon="add-new-solid"
+      icon="add-new"
       trigger={trigger || <Button {...buttonProps} as="button" />}
       mainBtnTitle="Create"
       mainAction={handleCreate}