+++ src/lib/components/select-contract/SelectContractAdmin.tsx
@@ -78,11 +78,7 @@ export const SelectContractAdmin = ({
         <DrawerOverlay />
         <DrawerContent h="80%">
           <DrawerHeader>
-            <CustomIcon
-              name="contract-address-solid"
-              boxSize={5}
-              color="gray.600"
-            />
+            <CustomIcon name="contract-address" boxSize={5} color="gray.600" />
             <Heading as="h5" variant="h5">
               Select contract which you have permission
             </Heading>