+++ CHANGELOG.md
@@ -37,15 +37,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+
 ### Features
 
+- [#1081](https://github.com/alleslabs/celatone-frontend/pull/1081) Add service APIs for move verify module
+- [#1082](https://github.com/alleslabs/celatone-frontend/pull/1082) Replace CONFIG_CHAINS by fetching the chains from the API
+- [#1077](https://github.com/alleslabs/celatone-frontend/pull/1077) Add download minitia custom JSON
+- [#1074](https://github.com/alleslabs/celatone-frontend/pull/1074) Add custom minitia upload JSON page
 - [#1002](https://github.com/alleslabs/celatone-frontend/pull/1002) Add contract verification modal
 - [#1060](https://github.com/alleslabs/celatone-frontend/pull/1060) Handle wasm verify info on various pages e.g. deploy/migrate flow pages, code/contract pages
 - [#1083](https://github.com/alleslabs/celatone-frontend/pull/1083) Add wasm verify badges on public pages and contract list pages
 
 ### Improvements
 
+- [#1034](https://github.com/alleslabs/celatone-frontend/pull/1034) Clean up custom icon file
+- [#1054](https://github.com/alleslabs/celatone-frontend/pull/1054) Improve main search bar to show both address and address as username
+
 ### Bug fixes
+
+- [#1075](https://github.com/alleslabs/celatone-frontend/pull/1075) Fix custom minitia minor change
 
 ## v1.7.3